refactor: `SpaceBlobAddHandler` is a `testutil`
